requirements.txt
openai
srsly
python-telegram-bot